优化日志

+++ src/main/java/com/sonnx/blog/dao/LogDao.java
@@ -28,7 +28,7 @@ public interface LogDao {
      * @param id
      * @return
      */
-    int deleteByPrimaryKey(Integer id);
+    Integer deleteByPrimaryKey(Long id);
 
     /**
      * @param record
